Plugin registration starting to become sensical. Session tests added.
Still need a larger startup config strategy, but will address that in
future work.
